Split widget: signals panel
