+++ extension/devtools/app/lib/main.dart
@@ -61,6 +61,7 @@ part 'widgets/batch_row.dart';
 part 'widgets/timeline_list.dart';
 part 'widgets/timeline_event_row.dart';
 part 'widgets/overview_panel.dart';
+part 'widgets/signals_panel.dart';
 part 'widgets/panel_placeholder.dart';
 part 'widgets/panel_scroll_view.dart';
 part 'widgets/panel_state_cards.dart';
@@ -215,90 +216,6 @@ class OrefPalette {
   static const Color pink = Color(0xFFFF71C6);
   static const Color deepBlue = Color(0xFF0C141C);
   static const Color lightBlue = Color(0xFFE7F3FF);
-}
-
-class _SignalsPanel extends StatelessWidget {
-  const _SignalsPanel();
-
-  @override
-  Widget build(BuildContext context) {
-    final state = _useSignalsPanelState(context);
-
-    return oref.SignalBuilder(
-      builder: (context) {
-        return _ConnectionGuard(
-          child: _PanelScrollView(
-            child: LayoutBuilder(
-              builder: (context, constraints) {
-                final controller = OrefDevToolsScope.of(context);
-                final entries = _samplesByKind(
-                  controller.snapshot?.samples ?? const <Sample>[],
-                  'signal',
-                );
-                final isSplit = constraints.maxWidth >= 980;
-                final filtered = state.filter(entries);
-                final selected = entries.firstWhereOrNull(
-                  (entry) => entry.id == state.selectedId(),
-                );
-                final list = _SignalList(
-                  entries: filtered,
-                  selectedId: selected?.id,
-                  isCompact: !isSplit,
-                  sortKey: state.sortKey(),
-                  sortAscending: state.sortAscending(),
-                  onSortName: () => state.toggleSort(_SortKey.name),
-                  onSortUpdated: () => state.toggleSort(_SortKey.updated),
-                  onSelect: (entry) => state.toggleSelection(entry.id),
-                );
-                final detail = _SignalDetail(entry: selected);
-
-                return Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    _SignalsHeader(
-                      controller: state.searchController,
-                      selectedFilter: state.statusFilter(),
-                      onFilterChange: (value) => state.statusFilter.set(value),
-                      totalCount: entries.length,
-                      filteredCount: filtered.length,
-                      onExport: () => _exportData(
-                        context,
-                        'signals',
-                        filtered.map((entry) => entry.toJson()).toList(),
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    if (isSplit)
-                      selected == null
-                          ? list
-                          : Row(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                Expanded(flex: 3, child: list),
-                                const SizedBox(width: 20),
-                                SizedBox(width: 320, child: detail),
-                              ],
-                            )
-                    else
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          list,
-                          if (selected != null) ...[
-                            const SizedBox(height: 16),
-                            detail,
-                          ],
-                        ],
-                      ),
-                  ],
-                );
-              },
-            ),
-          ),
-        );
-      },
-    );
-  }
 }
 
 class _ComputedPanel extends StatelessWidget {
